fix warning in MediaEngine

+++ Source/MediaEngine/Include/Core/Timestep.h
@@ -7,16 +7,16 @@ class Timestep
 public:
     Timestep() = default;
 
-    Timestep(float timeInSec)
+    Timestep(double timeInSec)
         : m_TimeInSec(timeInSec)
     {
     }
 
-    float GetSeconds() const { return m_TimeInSec; }
+    double GetSeconds() const { return m_TimeInSec; }
 
-    float GetMilliseconds() const { return m_TimeInSec * 1000; }
+    double GetMilliseconds() const { return m_TimeInSec * 1000; }
 
 private:
-    float m_TimeInSec = 0.f;
+    double m_TimeInSec = 0.f;
 };
 }  //namespace ME
